Fetch data mockup succesful
