+++ src/pages/signUpPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import GithubOAuth from "../components/githubOAuth";
 import { Link, useNavigate } from "react-router-dom";
-import { setCurrentUser, setErrorMessage } from "../store/userReducer";
+import { setCurrentUser, setErrorMessage, setAllPosts} from "../store/userReducer";
 import { useSelector, useDispatch } from "react-redux";
 
 
@@ -35,6 +35,67 @@ function SignUpPage () {
     }
 
   }
+
+  const mockData = [{
+    date: 5/23/2023,
+    userId: 'drake',
+    postTitle: 'I wish i was jamaican',
+    postBody: `Now I got a house in LA Now I got a bigger pool than Ye And look man, Ye's pool is nice Mine's just big is what I'm saying`,
+    postTag: 'default',
+    numLikes: 1,
+  },
+  {
+  date: 5/23/2023,
+  userId: 'wayne',
+  postTitle: 'I wish i was french',
+  postBody: `just testing this`,
+  postTag: 'default',
+  numLikes: 1,
+}]
+ 
+  
+
+  function getPosts () {
+    fetch('/')
+    .then(response => response.json())
+    .then(posts => {
+      
+      for (let i = 0; i < posts.length; i++) {
+          dispatch(setAllPosts(
+            [<div>
+              <h1>{posts[i].postTitle}</h1>
+              <p>{posts[i].postBody}</p>
+              <div>{posts[i].postTag}</div>
+              <div>{posts[i].numLikes}</div>
+              <div>{posts[i].numComments}</div>
+            </div>]
+          ))
+      }
+      
+    })
+    .catch(err => {
+      console.log('There was an error loading posts', err)
+    })
+  }
+
+
+  function mockGet (mockData) {
+    for (let i = 0; i < mockData.length; i++) {
+        dispatch(setAllPosts(
+          [<div>
+            <h1>{mockData[i].postTitle}</h1>
+            <p>{mockData[i].postBody}</p>
+            <div>{mockData[i].postTag}</div>
+            <div>{mockData[i].numLikes}</div>
+          </div> ]
+        ))
+      
+     
+    }
+   
+  }
+
+ 
 
   
  return (
@@ -71,7 +132,8 @@ function SignUpPage () {
           document.querySelector('#lastName').value,
           document.querySelector('#username').value,
           document.querySelector('#password').value)
-        }}>Sign Up</button>
+          mockGet(mockData)
+        }} m >Sign Up</button>
     </div>
 
     <GithubOAuth/>
